% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardFunctions.R
\name{addGoogleTracker}
\alias{addGoogleTracker}
\title{add google tracking code}
\usage{
addGoogleTracker(
  htmlPath,
  tag = readLines(file.path(webDirectory, "gatag.html"))
)
}
\arguments{
\item{htmlPath}{path to html file}

\item{tag}{tag code, default is readLines(file.path(webDirectory,"gatag.html"))}
}
\value{

}
\description{
add google tracking code
}
