% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addToDatabasee.R
\name{addLipdToDatabase}
\alias{addLipdToDatabase}
\title{Add or update a LiPD file in the database}
\usage{
addLipdToDatabase(
  L,
  dbPath = "/Users/nicholas/Dropbox/lipdverse/database/",
  standardize = FALSE
)
}
\arguments{
\item{dbPath}{local path to the database}
}
\description{
Add or update a LiPD file in the database
}
