% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardFunctions.R
\name{addLogoLink}
\alias{addLogoLink}
\title{add logo link}
\usage{
addLogoLink(htmlPath, link = "http://lipdverse.org")
}
\arguments{
\item{htmlPath}{path to html file}

\item{tag}{tag code, default is readLines(file.path(webDirectory,"gatag.html"))}
}
\value{

}
\description{
add logo link
}
