% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightlyUpdateDrake.R
\name{buildParams}
\alias{buildParams}
\title{Build parameters}
\usage{
buildParams(
  project,
  lipdDir,
  webDirectory,
  qcId,
  lastUpdateId,
  versionMetaId = "1OHD7PXEQ_5Lq6GxtzYvPA76bpQvN1_eYoFR0X80FIrY",
  googEmail = NULL,
  updateWebpages = TRUE,
  standardizeTerms = TRUE,
  ageOrYear = "age",
  restrictWebpagesToCompilation = TRUE,
  serialize = TRUE,
  projVersion = NA
)
}
\arguments{
\item{project}{project name}

\item{lipdDir}{authority directory for a lipd file}

\item{webDirectory}{directory for webserver}

\item{qcId}{google sheets ID for the qc sheet}

\item{lastUpdateId}{google sheets ID for the last version}

\item{googEmail}{google user ID}

\item{updateWebpages}{update lipdverse webpages (default = TRUE). Usually TRUE unless troubleshooting.}
}
\description{
Build parameters
}
