% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightlyUpdateDrake.R
\name{changeloggingAndUpdating}
\alias{changeloggingAndUpdating}
\title{Log changes and update}
\usage{
changeloggingAndUpdating(params, data)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
Log changes and update
}
