% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardizedTerminology.R
\name{checkBaseConflictsQcSheet}
\alias{checkBaseConflictsQcSheet}
\title{Check for conflicts in the qc sheet}
\usage{
checkBaseConflictsQcSheet(qc)
}
\arguments{
\item{qc}{a google qc sheet}
}
\value{
qc a (potentially updated) google qc sheet
}
\description{
Check for conflicts in the qc sheet
}
