% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightlyUpdateDrake.R
\name{checkIfUpdateNeeded}
\alias{checkIfUpdateNeeded}
\title{Check if an update is needed}
\usage{
checkIfUpdateNeeded(params)
}
\arguments{
\item{params}{}
}
\value{

}
\description{
Check if an update is needed
}
