% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/references.R
\name{createBib}
\alias{createBib}
\title{Create a RefManageR Bibentry for all the references in a multiLipd object}
\usage{
createBib(D)
}
\arguments{
\item{D}{multilipd object}
}
\value{
RefManageR Bibentry object
}
\description{
Create a RefManageR Bibentry for all the references in a multiLipd object
}
