% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardFunctions.R
\name{createDashboardRmd}
\alias{createDashboardRmd}
\title{Create an rmarkdown dashboard for a LiPD TS object}
\usage{
createDashboardRmd(thisTS, i, project, chronTS = NA)
}
\arguments{
\item{thisTS}{Lipd TSobject}

\item{i}{index to use for mapping wrt a larger TS}

\item{project}{name of the project}

\item{chronTS}{optional: LiPD chron TS object to include (default = NA)}
}
\value{
an Rmd file
}
\description{
Create an rmarkdown dashboard for a LiPD TS object
}
