% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addToDatabasee.R
\name{createDatabaseReference}
\alias{createDatabaseReference}
\title{create a reference table for the database}
\usage{
createDatabaseReference(D)
}
\arguments{
\item{D}{}
}
\value{

}
\description{
create a reference table for the database
}
