% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFileIndex.R
\name{createFileIndex}
\alias{createFileIndex}
\title{create a simple page that lists files in the directory}
\usage{
createFileIndex(thisDir, webDirectory = "~/Dropbox/lipdverse/html/")
}
\arguments{
\item{thisDir}{what directory to index?}

\item{webDirectory}{where's the main webdirectory}
}
\value{
writes an index html and Rmd fil
}
\description{
create a simple page that lists files in the directory
}
