% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipdScrubber.R
\name{createInterpretationGroupDirections}
\alias{createInterpretationGroupDirections}
\title{create directions for groups}
\usage{
createInterpretationGroupDirections(TS, groupFrom, groupInto)
}
\arguments{
\item{groupInto}{}
}
\description{
create directions for groups
}
