% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{createInventory}
\alias{createInventory}
\title{Create an inventory data.frame}
\usage{
createInventory(D)
}
\arguments{
\item{D}{a multi-lipd object}
}
\value{
a data.frame
}
\description{
Create an inventory data.frame
}
