% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{createInventoryGoogle}
\alias{createInventoryGoogle}
\title{Create a new inventory file on google drive}
\usage{
createInventoryGoogle(lipdDir)
}
\arguments{
\item{lipdDir}{the lipd directory of the database}

\item{newInv}{a inventory data frame}
}
\description{
Create a new inventory file on google drive
}
