% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcSheet.R
\name{createNewProject}
\alias{createNewProject}
\title{Create the files for a new project}
\usage{
createNewProject(
  templateID = "1JEm791Nhd4fUuyqece51CSlbR2A2I-pf8B0kFgwynug",
  project = "newProject",
  versionMetaId = "1OHD7PXEQ_5Lq6GxtzYvPA76bpQvN1_eYoFR0X80FIrY",
  googEmail = "nick.mckay2@gmail.com"
)
}
\arguments{
\item{googEmail}{}
}
\description{
Create the files for a new project
}
