% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcSheet.R
\name{createNewQCSheet}
\alias{createNewQCSheet}
\title{Create a new QC spreadsheet on google drive}
\usage{
createNewQCSheet(qcdf, qcName)
}
\arguments{
\item{qcdf}{QC Data.frame to upload}

\item{qcName}{Name of the qc file}
}
\description{
Create a new QC spreadsheet on google drive
}
