% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createProject.R
\name{createProjectDashboards}
\alias{createProjectDashboards}
\title{Create dahsboards for a project}
\usage{
createProjectDashboards(D, TS, webDirectory, project, version)
}
\arguments{
\item{D}{list of lipd files}

\item{TS}{lipd timeseries object}

\item{webDirectory}{website directory}

\item{project}{project name}

\item{version}{project version}

\item{currentVersion}{(TRUE or FALSE) copy to to currentVersion folder?}
}
\description{
Create dahsboards for a project
}
