% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardFunctions.R
\name{createProjectRmd}
\alias{createProjectRmd}
\title{Create an overview RMarkdown file for a project}
\usage{
createProjectRmd(project)
}
\arguments{
\item{project}{Name of the project}
}
\value{
Rmd file
}
\description{
Create an overview RMarkdown file for a project
}
