% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightlyUpdateDrake.R
\name{createProjectWebpages}
\alias{createProjectWebpages}
\title{Create lipdverse pages for this version of the project}
\usage{
createProjectWebpages(params, data)
}
\arguments{
\item{data}{}
}
\description{
Create lipdverse pages for this version of the project
}
