% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightlyUpdateDrake.R
\name{createQcFromFile}
\alias{createQcFromFile}
\title{Create QC sheet from data}
\usage{
createQcFromFile(params, data)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
Create QC sheet from data
}
