% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryCsv.R
\name{createQueryCsv}
\alias{createQueryCsv}
\title{Update the lipdverse query sheet with updated data}
\usage{
createQueryCsv(D)
}
\arguments{
\item{D}{a multiLipd object}
}
\value{
a data.frame for querying
}
\description{
Update the lipdverse query sheet with updated data
}
