% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipdScrubber.R
\name{createVectorsForGroups}
\alias{createVectorsForGroups}
\title{Create vectors for lumping}
\usage{
createVectorsForGroups(TS, groupFrom, groupInto)
}
\arguments{
\item{TS}{TS object}

\item{groupFrom}{name of TS variable to group}

\item{groupInto}{name of group to put into TS}
}
\value{
updated TS object
}
\description{
Create vectors for lumping
}
