% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightlyUpdateDrake.R
\name{createWebpages}
\alias{createWebpages}
\title{Create lipdversePages old framework}
\usage{
createWebpages(params, data)
}
\arguments{
\item{data}{}
}
\description{
Create lipdversePages old framework
}
