% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryCsv.R
\name{derivedMetadata}
\alias{derivedMetadata}
\title{Calculate derived metadata from a tibble ts}
\usage{
derivedMetadata(age, year, ...)
}
\arguments{
\item{...}{}
}
\description{
Calculate derived metadata from a tibble ts
}
