% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createProject.R
\name{directoryMD5}
\alias{directoryMD5}
\title{Calculate MD5 for a zipfile of a directory}
\usage{
directoryMD5(thisDir, pattern = "*.lpd")
}
\arguments{
\item{pattern}{}
}
\value{
MD5 string
}
\description{
Calculate MD5 for a zipfile of a directory
}
