% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipdScrubber.R
\name{fixExcelIssues}
\alias{fixExcelIssues}
\title{Fix issues from files imported from excel for climate 12k}
\usage{
fixExcelIssues(L)
}
\arguments{
\item{L}{}
}
\value{
L
}
\description{
Fix issues from files imported from excel for climate 12k
}
