% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipdScrubber.R
\name{fixKiloyearsTsChron}
\alias{fixKiloyearsTsChron}
\title{Convert kiloyears in a chron TS object}
\usage{
fixKiloyearsTsChron(TS)
}
\arguments{
\item{TS}{}
}
\value{
TS
}
\description{
Convert kiloyears in a chron TS object
}
