% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagUpdate.R
\name{flagUpdate}
\alias{flagUpdate}
\title{Turn on update flag}
\usage{
flagUpdate(compilation)
}
\arguments{
\item{compilation}{what compilation is updating?}
}
\value{
updates status website
}
\description{
Turn on update flag
}
