% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcSheet.R
\name{getGoogleQCSheet}
\alias{getGoogleQCSheet}
\title{get google qc sheet}
\usage{
getGoogleQCSheet(qcSheetId)
}
\arguments{
\item{qcSheetId}{Google sheet key}
}
\value{
qcs qc data frame
}
\description{
returns a qc sheet. Also saves a csv file called "googleQC.csv"
}
