% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipdScrubber.R
\name{getInterpretationGroupDirections}
\alias{getInterpretationGroupDirections}
\title{figure out  interpretation group directions}
\usage{
getInterpretationGroupDirections(vec)
}
\value{

}
\description{
figure out  interpretation group directions
}
