% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{getInventory}
\alias{getInventory}
\title{Get the lipdverse inventory}
\usage{
getInventory(lipdDir, googEmail)
}
\arguments{
\item{lipdDir}{the lipd directory of the database}
}
\value{
a data.frame
}
\description{
Get the lipdverse inventory
}
