% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightlyUpdate.R, R/nightlyUpdateDrake.R
\name{googleDriveUpdateTime}
\alias{googleDriveUpdateTime}
\title{Get google drive file update time}
\usage{
googleDriveUpdateTime(googId, tzone = "UTC")

googleDriveUpdateTime(googId, tzone = "UTC")
}
\arguments{
\item{googId}{Google drive file key}

\item{tzone}{timezone}
}
\value{



}
\description{
Get google drive file update time

Get google drive file update time
}
