% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcSheet.R
\name{hasDepth}
\alias{hasDepth}
\title{does the paleodata have depth?}
\usage{
hasDepth(L, dnames = c("depth"))
}
\arguments{
\item{dnames}{names to use to look for depth}
}
\description{
does the paleodata have depth?
}
