% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightlyUpdate.R
\name{lastVersion}
\alias{lastVersion}
\title{Get the most recent version of the compilation (before updating)}
\usage{
lastVersion(project,
  versionMetaId = "1OHD7PXEQ_5Lq6GxtzYvPA76bpQvN1_eYoFR0X80FIrY",
  googEmail = NULL)
}
\arguments{
\item{project}{project name}

\item{versionMetaId}{ID of the versioning qc sheet}

\item{googEmail}{google user ID}

\item{udsn}{a vector of dataset names in the project}
}
\value{
the new version string
}
\description{
Gets the last version of the database (before updating)
}
