% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcSheet.R
\name{nGoodAges}
\alias{nGoodAges}
\title{Get the number of good 14C ages in a Lipd file}
\usage{
nGoodAges(L, maxAge = 12000, c14names = c("age"))
}
\arguments{
\item{maxAge}{}
}
\value{

}
\description{
Get the number of good 14C ages in a Lipd file
}
