% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcSheet.R
\name{nUniqueAges}
\alias{nUniqueAges}
\title{Get the number of unique 14C ages in a Lipd file}
\usage{
nUniqueAges(L, maxAge = 12000, c14names = c("age14C", "x14Cage", "x14C",
  "DateBP"))
}
\arguments{
\item{maxAge}{}
}
\description{
Get the number of unique 14C ages in a Lipd file
}
