% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardFunctions.R
\name{plotCol}
\alias{plotCol}
\title{Plot a TS column as a in a lipd dashboard}
\usage{
plotCol(thisTS, ind, timeCol = NA)
}
\arguments{
\item{thisTS}{the ts to plot from}

\item{ind}{which index in the TS to plot}

\item{timeCol}{specify which variable is the x-axis (default = NA, which makes a reasonable guess)}
}
\value{
a graph
}
\description{
Plot a TS column as a in a lipd dashboard
}
