% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googlewriterUntitled.R
\name{read_sheet_retry}
\alias{read_sheet_retry}
\title{write google sheet with many attempts}
\usage{
read_sheet_retry(ss = NULL, sheet = NULL, ntries = 20, timeout = 1200, ...)
}
\arguments{
\item{ss}{ss}

\item{sheet}{sheet}

\item{ntries}{number of tries}

\item{timeout}{seconds before timeout. NA estimates from size.}

\item{data}{data.frame}
}
\description{
write google sheet with many attempts
}
