% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipdScrubber.R
\name{removeEmptyInterpretationsFromTs}
\alias{removeEmptyInterpretationsFromTs}
\title{remove empty interpretations from TS instance}
\usage{
removeEmptyInterpretationsFromTs(tsi)
}
\arguments{
\item{tsi}{}
}
\value{
tsi
}
\description{
remove empty interpretations from TS instance
}
