% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipdScrubber.R
\name{removeEmptyTables}
\alias{removeEmptyTables}
\title{Remove empty tables}
\usage{
removeEmptyTables(L, pc = "all")
}
\arguments{
\item{L}{a Lipd file}

\item{pc}{paleo or chron tables? (default= "all)}
}
\value{
a list of data.frames
}
\description{
Remove empty tables
}
