% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipdScrubber.R
\name{removeFakeConflicts}
\alias{removeFakeConflicts}
\title{remove artificial conflicts after merging}
\usage{
removeFakeConflicts(string)
}
\arguments{
\item{string}{a string}
}
\value{
a string with artificial conflicts removed
}
\description{
remove artificial conflicts after merging
}
