% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipdScrubber.R
\name{removeFakeConflictsCol}
\alias{removeFakeConflictsCol}
\title{remove artificial conflicts for a column}
\usage{
removeFakeConflictsCol(col)
}
\arguments{
\item{col}{}
}
\value{
col
}
\description{
remove artificial conflicts for a column
}
