% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonld.R
\name{removeJsonldFromHtml}
\alias{removeJsonldFromHtml}
\title{Remove json-ld tage from html}
\usage{
removeJsonldFromHtml(dataSetName, webdir)
}
\arguments{
\item{webdir}{}
}
\description{
Remove json-ld tage from html
}
