% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipdScrubber.R
\name{removeVariable}
\alias{removeVariable}
\title{remove all instances of a variable from a TS}
\usage{
removeVariable(TS, variable)
}
\arguments{
\item{variable}{}
}
\description{
remove all instances of a variable from a TS
}
