% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/romanize.R
\name{replaceSpecialCharacters}
\alias{replaceSpecialCharacters}
\title{Romanize special characters}
\usage{
replaceSpecialCharacters(toReplace, rosetta)
}
\arguments{
\item{toReplace}{a string, or vector of strings, to replace.}

\item{rosetta}{A data.frame that translate special characters, from rosettaStone()}
}
\value{
romanized strings
}
\description{
Romanize special characters
}
