% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipdScrubber.R
\name{resolveDumbConflicts}
\alias{resolveDumbConflicts}
\title{Fixes conflicts that shouldn't exist}
\usage{
resolveDumbConflicts(qc)
}
\arguments{
\item{qc}{}
}
\value{
qc
}
\description{
Fixes conflicts that shouldn't exist
}
