% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipdScrubber.R
\name{standardizeChronVariableNames}
\alias{standardizeChronVariableNames}
\title{standardize variableNames in a file}
\usage{
standardizeChronVariableNames(L)
}
\value{
L
}
\description{
standardizes the variable names in the chronData measurement tables.
}
