% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipdScrubber.R
\name{standardizeTsValues}
\alias{standardizeTsValues}
\title{Standardize keys in a TS}
\usage{
standardizeTsValues(TS,
  renamingDirectoryId = "1b-4arcNxxGsArCM6XfjW6y6yAp85BMlF-Gn8au4Zygg")
}
\description{
renames keys based on a directory of google spreadsheets
}
