% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipdScrubber.R
\name{standardizeValuesInValues}
\alias{standardizeValuesInValues}
\title{standardize values WITHIN a paleo(chron)Data_values field}
\usage{
standardizeValuesInValues(TS, googId)
}
\value{
TS
}
\description{
Importantly, this will run on every entry you give it, so you probably want to filter first.
}
