% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightlyUpdate.R
\name{tickVersion}
\alias{tickVersion}
\title{Title}
\usage{
tickVersion(project, udsn,
  versionMetaId = "1OHD7PXEQ_5Lq6GxtzYvPA76bpQvN1_eYoFR0X80FIrY")
}
\arguments{
\item{project}{project name}

\item{udsn}{a vector of dataset names in the project}

\item{versionMetaId}{ID of the versioning qc sheet}
}
\value{
the new version string
}
\description{
Ticks the version of a database for you. Assumes that a change is necessary.
}
