% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetId.R
\name{updateDatasetIdDereferencer}
\alias{updateDatasetIdDereferencer}
\title{update dataset ID database}
\usage{
updateDatasetIdDereferencer(
  D,
  compilation,
  version,
  dateUpdated = lubridate::today()
)
}
\arguments{
\item{dateUpdated}{}
}
\value{
a string describing the result
}
\description{
update dataset ID database
}
