% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{updateFromQC}
\alias{updateFromQC}
\title{update a split Timeseries object from a google QC sheet}
\usage{
updateFromQC(sTS, qcSheetId)
}
\arguments{
\item{sTS}{a split timeseries object}

\item{qcSheetId}{a google drive id}
}
\value{
an updated sTS
}
\description{
update a split Timeseries object from a google QC sheet
}
