% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{updateInventory}
\alias{updateInventory}
\title{Update the lipdverse inventory}
\usage{
updateInventory(newInv, lipdDir)
}
\arguments{
\item{newInv}{the updated inventory data frame}

\item{lipdDir}{the lipd directory of the database}
}
\description{
Update the lipdverse inventory
}
