% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightlyUpdate.R
\name{updateNeeded}
\alias{updateNeeded}
\title{Check to see if a project needs to be updated}
\usage{
updateNeeded(project, webDirectory, lipdDir, qcId,
  versionMetaId = "1OHD7PXEQ_5Lq6GxtzYvPA76bpQvN1_eYoFR0X80FIrY")
}
\value{
TRUE or FALSE
}
\description{
Check to see if a project needs to be updated
}
