% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightlyUpdate.R
\name{updateProject}
\alias{updateProject}
\title{Do a complete update to a project}
\usage{
updateProject(project, lipdDir, webDirectory, qcId, lastUpdateId,
  versionMetaId = "1OHD7PXEQ_5Lq6GxtzYvPA76bpQvN1_eYoFR0X80FIrY",
  googEmail = NULL, updateWebpages = TRUE)
}
\arguments{
\item{project}{project name}

\item{lipdDir}{authority directory for a lipd file}

\item{webDirectory}{directory for webserver}

\item{qcId}{google sheets ID for the qc sheet}

\item{lastUpdateId}{google sheets ID for the last version}

\item{googEmail}{google user ID}

\item{updateWebpages}{update lipdverse webpages (default = TRUE). Usually TRUE unless troubleshooting.}
}
\description{
Do a complete update to a project
}
