% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryCsv.R
\name{updateQueryCsv}
\alias{updateQueryCsv}
\title{Update the lipdverse query sheet with updated data}
\usage{
updateQueryCsv(D)
}
\arguments{
\item{D}{a multiLipd object}
}
\description{
Update the lipdverse query sheet with updated data
}
