% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googlewriterUntitled.R
\name{write_sheet_retry}
\alias{write_sheet_retry}
\title{write google sheet with many attempts}
\usage{
write_sheet_retry(data, ss = NULL, sheet = NULL, ntries = 50, timeout = NA)
}
\arguments{
\item{data}{data.frame}

\item{ss}{ss}

\item{sheet}{sheet}

\item{ntries}{number of tries}

\item{timeout}{seconds before timeout. NA estimates from size.}
}
\description{
write google sheet with many attempts
}
