% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardFunctions.R
\name{writeCollapsibleChunks}
\alias{writeCollapsibleChunks}
\title{Build collapsible metadata for RMarkdown from a lipd TS object}
\usage{
writeCollapsibleChunks(
  thisRmd,
  thisTS = thisTS,
  name = "pub",
  vars = c("author", "citeKey", "journal", "volume", "pages", "pubYear", "title",
    "DOI"),
  tsi = 1,
  indent = 0,
  forceName = NULL,
  dontClose = FALSE,
  open = FALSE
)
}
\arguments{
\item{thisRmd}{Rmd file to start with}

\item{thisTS}{TS object}

\item{name}{name of the metadata chunk}

\item{vars}{name of the variables to include in the list}

\item{tsi}{index for the TS object (default = 1)}

\item{indent}{indentation level for the chunk (default = 0)}

\item{forceName}{define a name for the chunk that doesn't match the name in the TS (default = NULL)}

\item{dontClose}{optionally, don't terminate the collapsible data, allowing you to add more to the list (default = FALSE)}

\item{open}{should the collapsible list stay open? (default = FALSE)}
}
\value{
an updated Rmarkdown file
}
\description{
Build collapsible metadata for RMarkdown from a lipd TS object
}
