% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{createQCdataFrame}
\alias{createQCdataFrame}
\title{Title}
\usage{
createQCdataFrame(sTS, templateId, to.omit = c("depth", "age", "year"),
  to.omit.specific = c("yr"))
}
\value{
a data.frame QC sheet
}
\description{
Title
}
