% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcSheet.R
\name{createQCdataFrame}
\alias{createQCdataFrame}
\title{Create QC data.frame}
\usage{
createQCdataFrame(
  sTS,
  templateId,
  to.omit = c("age", "year"),
  to.omit.specific = c("depth", "yr"),
  ageOrYear = "age",
  compilationName = NA,
  compVersion = NA
)
}
\value{
a data.frame QC sheet
}
\description{
Create QC data.frame
}
