% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcSheet.R
\name{getMostRecentInCompilationsTs}
\alias{getMostRecentInCompilationsTs}
\title{Get most recent compilations and versions}
\usage{
getMostRecentInCompilationsTs(TS)
}
\value{
a character vector of compilations and versions
}
\description{
Get most recent compilations and versions
}
