% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcSheet.R
\name{nUniqueAges}
\alias{nUniqueAges}
\title{Get the number of unique 14C ages in a Lipd file}
\usage{
nUniqueAges(L, maxAge = 12000, c14names = c("age"))
}
\arguments{
\item{maxAge}{}
}
\value{

}
\description{
Get the number of unique 14C ages in a Lipd file
}
