% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightlyUpdate.R, R/nightlyUpdateDrake.R
\name{tickVersion}
\alias{tickVersion}
\title{Title}
\usage{
tickVersion(
  project,
  qcIc,
  tsIc,
  versionMetaId = "1OHD7PXEQ_5Lq6GxtzYvPA76bpQvN1_eYoFR0X80FIrY",
  googEmail = NULL
)

tickVersion(
  project,
  qcIc,
  tsIc,
  versionMetaId = "1OHD7PXEQ_5Lq6GxtzYvPA76bpQvN1_eYoFR0X80FIrY",
  googEmail = NULL
)
}
\arguments{
\item{project}{project name}

\item{qcIc}{dataSetNames in this compilation from teh QC sheet}

\item{tsIc}{dataSetNames in the last compilation from the files}

\item{versionMetaId}{ID of the versioning qc sheet}

\item{googEmail}{google user ID}
}
\value{
the new version string

the new version string
}
\description{
Ticks the version of a database for you. Assumes that a change is necessary.

Ticks the version of a database for you. Assumes that a change is necessary.
}
