% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcSheet.R
\name{updateFromQC}
\alias{updateFromQC}
\title{update a split Timeseries object from a google QC sheet}
\usage{
updateFromQC(sTS, qcs, compilationName = "test", newVersion = "0.0.0")
}
\arguments{
\item{sTS}{a split timeseries object}

\item{qcs}{a QC data.frame}
}
\value{
an updated sTS
}
\description{
update a split Timeseries object from a google QC sheet
}
