% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightlyUpdate.R
\name{updateProject}
\alias{updateProject}
\title{Do a complete update to a project}
\usage{
updateProject(project, lipdDir, webDirectory, qcId)
}
\arguments{
\item{project}{project name}

\item{lipdDir}{authority directory for a lipd file}

\item{webDirectory}{directory for webserver}

\item{qcId}{google sheets ID for the qc sheet}
}
\description{
Do a complete update to a project
}
